renamed the integration and added github templates
