+++ .devcontainer/configuration.yaml
@@ -4,7 +4,7 @@ default_config:
 logger:
   default: warning
   logs:
-    custom_components.go_echarger: debug
+    custom_components.smart_energy_goecharger: debug
     custom_components.hacs: debug
     aiogithubapi: debug
 
@@ -13,7 +13,7 @@ sensor:
     display_options:
       - 'date_time'
 
-go_echarger:
+smart_energy_goecharger:
   chargers:
     - name: charger1
       host: http://localhost:4000
