+++ .dockerignore
@@ -4,7 +4,7 @@
 env
 custom_components/___pycache__
 custom_components/___init__.py
-custom_components/go_echarger/__pycache__
+custom_components/smart_energy_goecharger/__pycache__
 tests
 
 .coverage
