+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,24 @@
+---
+name: Bug report
+about: Create a report for a discovered issue.
+labels: bug
+author: ''
+
+---
+
+## Summary and context
+
+> Hint: A clear and concise description of what the bug is.
+
+## How to reproduce
+
+> Hint: Briefly (in points) describe how to replicate your issue.
+
+## Expected behavior
+
+> Hint: A clear and concise description of what you expected to happen.
+
+## Versions (please complete the following information):**
+
+ - smart_energy-goecharger
+ - Home Assistant
