+++ .github/pull_request_template.md
@@ -1,0 +1,20 @@
+---
+name: Pull request
+about: Create a pull request for a proposed change.
+author: ''
+
+---
+
+## Description
+
+> Hint: Briefly describe what has changed. Link to an issue if applicable.
+
+## How to test
+
+> Hint: Briefly (in points) describe how to test your changes.
+
+## Checklist
+
+- [ ] Change is covered by unit test (if it makes sense)
+- [ ] Changelog is updated
+- [ ] Integration version is increased (major/minor/bugfix)
