+++ Dockerfile
@@ -1,5 +1,5 @@
 FROM homeassistant/home-assistant:stable
 
-RUN python3 -m pip install "goechargerv2==0.2.2"
+RUN python3 -m pip install "smart-energy.goecharger-api==0.3.0"
 
 COPY configuration.yaml /config/configuration.yaml
