+++ Dockerfile.dev
@@ -1,6 +1,6 @@
 FROM homeassistant/home-assistant:stable
 
-RUN python3 -m pip install "goechargerv2==0.2.2"
+RUN python3 -m pip install "smart-energy.goecharger-api==0.3.0"
 
-COPY custom_components/go_echarger /config/custom_components/go_echarger
+COPY custom_components/smart_energy_goecharger /config/custom_components/smart_energy_goecharger
 COPY configuration.dev.yaml /config/configuration.yaml
