+++ configuration.yaml
@@ -4,7 +4,7 @@ default_config:
 logger:
   default: warning
   logs:
-    custom_components.go_echarger: debug
+    custom_components.smart_energy_goecharger: debug
     custom_components.hacs: debug
     aiogithubapi: debug
