--- custom_components/go_echarger/manifest.json
@@ -1,16 +1,0 @@
-{
-  "domain": "go_echarger",
-  "name": "go-e Charger Cloud",
-  "version": "0.2.4",
-  "documentation": "https://github.com/openkfw/homeassistant_goechargerv2",
-  "issue_tracker": "https://github.com/openkfw/homeassistant_goechargerv2/issues",
-  "iot_class": "local_polling",
-  "config_flow": true,
-  "dependencies": ["sensor", "switch", "button", "number", "select"],
-  "codeowners": [
-    "@openkfw"
-  ],
-  "requirements": [
-    "goechargerv2==0.2.2"
-  ]
-}
