+++ custom_components/smart_energy_goecharger/translations/en.json
