+++ custom_components/smart_energy_goecharger/translations/select.en.json
@@ -1,6 +1,6 @@
 {
     "state": {
-        "go_echarger__phase_switch_mode": {
+        "smart_energy_goecharger__phase_switch_mode": {
             "0": "Auto",
             "1": "1-phase",
             "2": "3-phase"
