+++ docs/dev.md
@@ -30,7 +30,7 @@ Run:
 ./start-dev.sh
 ```
 
-After few minutes, Home Assistant should be running and script should be in the watch mode. Whenever you change a file in the `custom_components/go_echarger` folder, it will restart the Home Assistant within a few seconds. Thus, you have a quick development feedback.
+After few minutes, Home Assistant should be running and script should be in the watch mode. Whenever you change a file in the `custom_components/smart_energy_goecharger` folder, it will restart the Home Assistant within a few seconds. Thus, you have a quick development feedback.
 
 1. Open the <http://127.0.0.1:8123> in a browser.
 2. Create an account.
@@ -49,7 +49,7 @@ In case you want to try HACS locally, run:
 2. Create an account - make sure that timezone is set correctly, otherwise it will fail to connect to the Github.
 3. In the left menu you should have HACS icon, click it.
 4. Click on `Integrations` -> click 3 dots top right corner -> click `Custom repositories`.
-5. In the dialog window, add `https://github.com/openkfw/homeassistant_goechargerv2` as a repository and select `Integration` as a category.
+5. In the dialog window, add `https://github.com/openkfw/smart_energy.goecharger` as a repository and select `Integration` as a category.
 6. Click `ADD`, wait for spinner to finish and close the dialog.
 7. Click `EXPLORE & DOWNLOAD REPOSITORIES` -> search for `go-e` -> select the `go-e Charger Cloud` -> wait and click `DOWNLOAD`.
 8. Go to Settings -> System -> click `RESTART` and wait few seconds.
@@ -114,7 +114,7 @@ pylint tests/**/*.py custom_components/**/*.py mock_api/**/*.py
 Formatting is done via [Black](https://black.readthedocs.io/en/stable/getting_started.html).
 
 ```
-black custom_components/go_echarger
+black custom_components/smart_energy_goecharger
 ```
 
 To have autoformatting in the VSCode, install the extension `ms-python.black-formatter`.
@@ -130,7 +130,7 @@ pytest
 pytest -o log_cli=true
 
 # code coverage
-pytest --durations=10 --cov-report term-missing --cov=custom_components.go_echarger tests
+pytest --durations=10 --cov-report term-missing --cov=custom_components.smart_energy_goecharger tests
 ```
 
 > __Note: In case you have issues with bcrypt circular import, run this:__
