+++ pyproject.toml
@@ -13,17 +13,17 @@ classifiers = [
   "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent",
 ]
-dependencies = ["goechargerv2"]
+dependencies = ["smart-energy.goecharger-api"]
 description = "This is an integration for the Home Assistant for the go-e Charger Cloud wallbox (API v2)."
 keywords = ["goecharger", "kfw", "homeassistant", "cloud"]
-name = "homeassistant_goechargerv2"
+name = "smart_energy_goecharger"
 readme = "README.md"
 requires-python = ">=3.7"
-version = "0.2.4"
+version = "0.3.0"
 
 [project.optional-dependencies]
 dev = ["black", "pylint", "pytest-homeassistant-custom-component", "Flask", "pre-commit"]
 
 [project.urls]
-"Bug Tracker" = "https://github.com/openkfw/homeassistant_goechargerv2/issues"
-"Homepage" = "https://github.com/openkfw/homeassistant_goechargerv2"
+"Bug Tracker" = "https://github.com/openkfw/smart_energy.goecharger/issues"
+"Homepage" = "https://github.com/openkfw/smart_energy.goecharger"
